Add groupTypes to complet accountTypes
